1. Removed DP code from CP 2. Removed legacy code regarding RT/RT6/MPLS messages 3. Bug Fixes 4. Added Tracing to Trace all CP2DP msgs recvd on DP

+++ datapath/Vrfs/dp_vrf.h
@@ -49,7 +49,7 @@ void
 dp_init_vrf_hashtable (hashtable_t **ht);
 
 dp_vrf_t *
-dp_look_up_vrf (hashtable_t *ht, uint8_t vrf_id);
+dp_look_up_vrf (hashtable_t *ht, int16_t vrf_id);
 
 void
 dp_insert_vrf (hashtable_t *ht, dp_vrf_t *vrf);
